[Refactor] main에서 home과 mypage 분리
