+++ app/src/main/java/com/doctor/yumyum/presentation/ui/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.doctor.yumyum.presentation.viewmodel
+package com.doctor.yumyum.presentation.ui.home
 
 import com.doctor.yumyum.common.base.BaseViewModel
