+++ app/src/main/java/com/doctor/yumyum/presentation/ui/main/MainActivity.kt
@@ -6,6 +6,8 @@ import androidx.lifecycle.ViewModelProvider
 import com.doctor.yumyum.R
 import com.doctor.yumyum.databinding.ActivityMainBinding
 import com.doctor.yumyum.common.base.BaseActivity
+import com.doctor.yumyum.presentation.ui.home.HomeFragment
+import com.doctor.yumyum.presentation.ui.mypage.MyPageFragment
 import com.doctor.yumyum.presentation.ui.write.WriteRecipeActivity
 
 class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
