+++ app/src/main/java/com/doctor/yumyum/presentation/ui/mypage/MyPageViewModel.kt
@@ -1,4 +1,4 @@
-package com.doctor.yumyum.presentation.viewmodel
+package com.doctor.yumyum.presentation.ui.mypage
 
 import com.doctor.yumyum.common.base.BaseViewModel
