+++ app/src/main/res/layout/fragment_my_page.xml
@@ -6,14 +6,14 @@
 
         <variable
             name="viewModel"
-            type="com.doctor.yumyum.presentation.viewmodel.MyPageViewModel" />
+            type="com.doctor.yumyum.presentation.ui.mypage.MyPageViewModel" />
 
     </data>
 
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.main.MyPageFragment">
+        tools:context=".presentation.ui.mypage.MyPageFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
